Update num_abbr.py
I'm only assuming that the scale is responsible for the position after decimal point of the download counter, my knowledge of Python is very low. However, I would like to suggest to set the scale to `1` for downloads in their thousands, because the current setting is not very precise (e.g. 1.6K is displayed as 2K.)